backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.2.0
sentence-transformers==2.2.2
chromadb==0.5.12
langchain==0.0.263
openai==1.0.0
pydantic==1.10.7
pytest==7.4.0
